backend/requirements.txt
django
djangorestframework
cryptocompare
django-cors-headers